+++ src/concurrency/built-in/Context.ts
@@ -1,5 +1,6 @@
 
 import * as puppeteer from 'puppeteer';
+import { BROWSER_TIMEOUT, timeoutExecute } from '../../util';
 
 import { ResourceData } from '../ConcurrencyImplementation';
 import SingleBrowserImplementation from '../SingleBrowserImplementation';
@@ -17,7 +18,7 @@ export default class Context extends SingleBrowserImplementation {
     }
 
     protected async freeResources(resources: ResourceData): Promise<void> {
-        await resources.context.close();
+        await timeoutExecute(BROWSER_TIMEOUT,  resources.context.close());
     }
 
 }